+++ src/Agent/NewRelic/Agent/Core/DataTransport/ConnectModel.cs
@@ -55,7 +55,7 @@ namespace NewRelic.Agent.Core.DataTransport
         public readonly IEnumerable<Label> Labels;
 
         [JsonProperty("settings")]
-        public readonly JavascriptAgentSettingsModel JavascriptAgentSettings;
+        public readonly ReportedConfiguration Configuration;
 
         [JsonProperty("metadata")]
         public readonly Dictionary<string, string> Metadata;
@@ -69,7 +69,7 @@ namespace NewRelic.Agent.Core.DataTransport
         [JsonProperty("security_policies", NullValueHandling = NullValueHandling.Ignore)]
         public readonly SecurityPoliciesSettingsModel SecurityPoliciesSettings;
 
-        public ConnectModel(int processId, string language, string displayHost, string hostName, IEnumerable<string> appNames, string agentVersion, long agentVersionTimestamp, SecuritySettingsModel securitySettings, bool highSecurityModeEnabled, string identifier, IEnumerable<Label> labels, JavascriptAgentSettingsModel javascriptAgentSettings, Dictionary<string, string> metadata, UtilizationSettingsModel utilizationSettings, Environment environment, SecurityPoliciesSettingsModel securityPoliciesSettings, EventHarvestConfigModel eventHarvestConfig)
+        public ConnectModel(int processId, string language, string displayHost, string hostName, IEnumerable<string> appNames, string agentVersion, long agentVersionTimestamp, SecuritySettingsModel securitySettings, bool highSecurityModeEnabled, string identifier, IEnumerable<Label> labels, Dictionary<string, string> metadata, UtilizationSettingsModel utilizationSettings, Environment environment, SecurityPoliciesSettingsModel securityPoliciesSettings, EventHarvestConfigModel eventHarvestConfig, ReportedConfiguration configuration)
         {
             ProcessId = processId;
             Language = language;
@@ -82,12 +82,12 @@ namespace NewRelic.Agent.Core.DataTransport
             HighSecurityModeEnabled = highSecurityModeEnabled;
             Identifier = identifier;
             Labels = labels;
-            JavascriptAgentSettings = javascriptAgentSettings;
             Metadata = metadata;
             UtilizationSettings = utilizationSettings;
             Environment = environment;
             SecurityPoliciesSettings = securityPoliciesSettings;
             EventHarvestConfig = eventHarvestConfig;
+            Configuration = configuration;
         }
     }
 
@@ -110,21 +110,6 @@ namespace NewRelic.Agent.Core.DataTransport
         public TransactionTraceSettingsModel(string recordSql)
         {
             RecordSql = recordSql;
-        }
-    }
-
-    public class JavascriptAgentSettingsModel
-    {
-        [JsonProperty("browser_monitoring.loader_debug")]
-        public readonly bool LoaderDebug;
-
-        [JsonProperty("browser_monitoring.loader")]
-        public readonly string Loader;
-
-        public JavascriptAgentSettingsModel(bool loaderDebug, string loader)
-        {
-            LoaderDebug = loaderDebug;
-            Loader = loader;
         }
     }