+++ tests/Agent/IntegrationTests/Applications/Owin3WebApi/Controllers/AsyncAwaitController.cs
@@ -46,6 +46,9 @@ namespace Owin3WebApi.Controllers
             await Task.Run(() => TaskRunBackgroundMethod());
             await Task.Factory.StartNew(TaskFactoryStartNewBackgroundMethod);
 
+            // Try to force this to be the slowest transaction
+            await Task.Delay(TimeSpan.FromSeconds(5));
+
             return "Worked";
         }