+++ tests/Agent/IntegrationTests/IntegrationTests/Logging/MetricsAndForwardingTests.cs
@@ -577,14 +577,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
-        public class Log4NetMetricsAndForwardingEnabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<ConsoleDynamicMethodFixtureCore21>
-        {
-            public Log4NetMetricsAndForwardingEnabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
-                : base(fixture, output, true, true, true, LoggingFramework.Log4net)
-            {
-            }
-        }
         #endregion
 
         #region Metrics and Forwarding Disabled
@@ -637,15 +629,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
         public class Log4NetMetricsAndForwardingDisabledTestsNetCore31Tests : MetricsAndForwardingTestsBase<ConsoleDynamicMethodFixtureCore31>
         {
             public Log4NetMetricsAndForwardingDisabledTestsNetCore31Tests(ConsoleDynamicMethodFixtureCore31 fixture, ITestOutputHelper output)
-                : base(fixture, output, false, false, true, LoggingFramework.Log4net)
-            {
-            }
-        }
-
-        [NetCoreTest]
-        public class Log4NetMetricsAndForwardingDisabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<ConsoleDynamicMethodFixtureCore21>
-        {
-            public Log4NetMetricsAndForwardingDisabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
                 : base(fixture, output, false, false, true, LoggingFramework.Log4net)
             {
             }
@@ -708,15 +691,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
-        public class Log4NetMetricsEnabledAndForwardingDisabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<ConsoleDynamicMethodFixtureCore21>
-        {
-            public Log4NetMetricsEnabledAndForwardingDisabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
-                : base(fixture, output, true, false, true, LoggingFramework.Log4net)
-            {
-            }
-        }
-
         #endregion
 
         #region Metrics Disabled, Forwarding Enabled
@@ -774,24 +748,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
-        public class Log4netMetricsDisabledAndForwardingEnabledTestsNetCore22Tests : MetricsAndForwardingTestsBase<ConsoleDynamicMethodFixtureCore22>
-        {
-            public Log4netMetricsDisabledAndForwardingEnabledTestsNetCore22Tests(ConsoleDynamicMethodFixtureCore22 fixture, ITestOutputHelper output)
-                : base(fixture, output, false, true, true, LoggingFramework.Log4net)
-            {
-            }
-        }
-
-        [NetCoreTest]
-        public class Log4netMetricsDisabledAndForwardingEnabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<ConsoleDynamicMethodFixtureCore21>
-        {
-            public Log4netMetricsDisabledAndForwardingEnabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
-                : base(fixture, output, false, true, true, LoggingFramework.Log4net)
-            {
-            }
-        }
-
         #endregion
     }
 
@@ -839,13 +795,13 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
+        [NetFrameworkTest]
         public class
-            MicrosoftLoggingMetricsAndForwardingEnabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
+        MicrosoftLoggingMetricsAndForwardingEnabledTestsFWLatestTests : MetricsAndForwardingTestsBase<
+        ConsoleDynamicMethodFixtureFWLatest>
         {
-            public MicrosoftLoggingMetricsAndForwardingEnabledTestsNetCore21Tests(
-                ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
+            public MicrosoftLoggingMetricsAndForwardingEnabledTestsFWLatestTests(
+                ConsoleDynamicMethodFixtureFWLatest fixture, ITestOutputHelper output)
                 : base(fixture, output, true, true, true, LoggingFramework.MicrosoftLogging)
             {
             }
@@ -891,13 +847,13 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
+        [NetFrameworkTest]
         public class
-            MicrosoftLoggingMetricsAndForwardingDisabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
+            MicrosoftLoggingMetricsAndForwardingDisabledTestsFWLatestTests : MetricsAndForwardingTestsBase<
+        ConsoleDynamicMethodFixtureFWLatest>
         {
-            public MicrosoftLoggingMetricsAndForwardingDisabledTestsNetCore21Tests(
-                ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
+            public MicrosoftLoggingMetricsAndForwardingDisabledTestsFWLatestTests(
+                ConsoleDynamicMethodFixtureFWLatest fixture, ITestOutputHelper output)
                 : base(fixture, output, false, false, true, LoggingFramework.MicrosoftLogging)
             {
             }
@@ -943,13 +899,13 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
+        [NetFrameworkTest]
         public class
-            MicrosoftLoggingMetricsEnabledAndForwardingDisabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
+            MicrosoftLoggingMetricsEnabledAndForwardingDisabledTestsFWLatestTests : MetricsAndForwardingTestsBase<
+        ConsoleDynamicMethodFixtureFWLatest>
         {
-            public MicrosoftLoggingMetricsEnabledAndForwardingDisabledTestsNetCore21Tests(
-                ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
+            public MicrosoftLoggingMetricsEnabledAndForwardingDisabledTestsFWLatestTests(
+                ConsoleDynamicMethodFixtureFWLatest fixture, ITestOutputHelper output)
                 : base(fixture, output, true, false, true, LoggingFramework.MicrosoftLogging)
             {
             }
@@ -995,25 +951,13 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
+        [NetFrameworkTest]
         public class
-            MicrosoftLoggingMetricsDisabledAndForwardingEnabledTestsNetCore22Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore22>
+            MicrosoftLoggingMetricsDisabledAndForwardingEnabledTestsFWLatestTests : MetricsAndForwardingTestsBase<
+                ConsoleDynamicMethodFixtureFWLatest>
         {
-            public MicrosoftLoggingMetricsDisabledAndForwardingEnabledTestsNetCore22Tests(
-                ConsoleDynamicMethodFixtureCore22 fixture, ITestOutputHelper output)
-                : base(fixture, output, false, true, true, LoggingFramework.MicrosoftLogging)
-            {
-            }
-        }
-
-        [NetCoreTest]
-        public class
-            MicrosoftLoggingMetricsDisabledAndForwardingEnabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
-        {
-            public MicrosoftLoggingMetricsDisabledAndForwardingEnabledTestsNetCore21Tests(
-                ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
+            public MicrosoftLoggingMetricsDisabledAndForwardingEnabledTestsFWLatestTests(
+                ConsoleDynamicMethodFixtureFWLatest fixture, ITestOutputHelper output)
                 : base(fixture, output, false, true, true, LoggingFramework.MicrosoftLogging)
             {
             }
@@ -1096,18 +1040,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
                 ConsoleDynamicMethodFixtureCore31>
         {
             public SerilogMetricsAndForwardingEnabledTestsNetCore31Tests(ConsoleDynamicMethodFixtureCore31 fixture,
-                ITestOutputHelper output)
-                : base(fixture, output, true, true, true, LoggingFramework.Serilog)
-            {
-            }
-        }
-
-        [NetCoreTest]
-        public class
-            SerilogMetricsAndForwardingEnabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
-        {
-            public SerilogMetricsAndForwardingEnabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture,
                 ITestOutputHelper output)
                 : base(fixture, output, true, true, true, LoggingFramework.Serilog)
             {
@@ -1202,18 +1134,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
-        public class
-            SerilogMetricsAndForwardingDisabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
-        {
-            public SerilogMetricsAndForwardingDisabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture,
-                ITestOutputHelper output)
-                : base(fixture, output, false, false, true, LoggingFramework.Serilog)
-            {
-            }
-        }
-
         #endregion
 
         #region Metrics Enabled, Forwarding Disabled
@@ -1290,18 +1210,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
-        public class
-            SerilogMetricsEnabledAndForwardingDisabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
-        {
-            public SerilogMetricsEnabledAndForwardingDisabledTestsNetCore21Tests(
-                ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
-                : base(fixture, output, true, false, true, LoggingFramework.Serilog)
-            {
-            }
-        }
-
         #endregion
 
         #region Metrics Disabled, Forwarding Enabled
@@ -1373,30 +1281,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
         {
             public SerilogMetricsDisabledAndForwardingEnabledTestsNetCore31Tests(
                 ConsoleDynamicMethodFixtureCore31 fixture, ITestOutputHelper output)
-                : base(fixture, output, false, true, true, LoggingFramework.Serilog)
-            {
-            }
-        }
-
-        [NetCoreTest]
-        public class
-            SerilogMetricsDisabledAndForwardingEnabledTestsNetCore22Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore22>
-        {
-            public SerilogMetricsDisabledAndForwardingEnabledTestsNetCore22Tests(
-                ConsoleDynamicMethodFixtureCore22 fixture, ITestOutputHelper output)
-                : base(fixture, output, false, true, true, LoggingFramework.Serilog)
-            {
-            }
-        }
-
-        [NetCoreTest]
-        public class
-            SerilogMetricsDisabledAndForwardingEnabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
-        {
-            public SerilogMetricsDisabledAndForwardingEnabledTestsNetCore21Tests(
-                ConsoleDynamicMethodFixtureCore21 fixture, ITestOutputHelper output)
                 : base(fixture, output, false, true, true, LoggingFramework.Serilog)
             {
             }
@@ -1485,18 +1369,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
-        public class
-            NLogMetricsAndForwardingEnabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
-        {
-            public NLogMetricsAndForwardingEnabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture,
-                ITestOutputHelper output)
-                : base(fixture, output, true, true, true, LoggingFramework.NLog)
-            {
-            }
-        }
-
         #endregion
 
         #region Metrics and Forwarding Disabled
@@ -1567,18 +1439,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
                 ConsoleDynamicMethodFixtureCore31>
         {
             public NLogMetricsAndForwardingDisabledTestsNetCore31Tests(ConsoleDynamicMethodFixtureCore31 fixture,
-                ITestOutputHelper output)
-                : base(fixture, output, false, false, true, LoggingFramework.NLog)
-            {
-            }
-        }
-
-        [NetCoreTest]
-        public class
-            NLogMetricsAndForwardingDisabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
-        {
-            public NLogMetricsAndForwardingDisabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture,
                 ITestOutputHelper output)
                 : base(fixture, output, false, false, true, LoggingFramework.NLog)
             {
@@ -1661,18 +1521,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
             }
         }
 
-        [NetCoreTest]
-        public class
-            NLogMetricsEnabledAndForwardingDisabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
-        {
-            public NLogMetricsEnabledAndForwardingDisabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture,
-                ITestOutputHelper output)
-                : base(fixture, output, true, false, true, LoggingFramework.NLog)
-            {
-            }
-        }
-
         #endregion
 
         #region Metrics Disabled, Forwarding Enabled
@@ -1743,30 +1591,6 @@ namespace NewRelic.Agent.IntegrationTests.Logging.MetricsAndForwarding
                 ConsoleDynamicMethodFixtureCore31>
         {
             public NLogMetricsDisabledAndForwardingEnabledTestsNetCore31Tests(ConsoleDynamicMethodFixtureCore31 fixture,
-                ITestOutputHelper output)
-                : base(fixture, output, false, true, true, LoggingFramework.NLog)
-            {
-            }
-        }
-
-        [NetCoreTest]
-        public class
-            NLogMetricsDisabledAndForwardingEnabledTestsNetCore22Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore22>
-        {
-            public NLogMetricsDisabledAndForwardingEnabledTestsNetCore22Tests(ConsoleDynamicMethodFixtureCore22 fixture,
-                ITestOutputHelper output)
-                : base(fixture, output, false, true, true, LoggingFramework.NLog)
-            {
-            }
-        }
-
-        [NetCoreTest]
-        public class
-            NLogMetricsDisabledAndForwardingEnabledTestsNetCore21Tests : MetricsAndForwardingTestsBase<
-                ConsoleDynamicMethodFixtureCore21>
-        {
-            public NLogMetricsDisabledAndForwardingEnabledTestsNetCore21Tests(ConsoleDynamicMethodFixtureCore21 fixture,
                 ITestOutputHelper output)
                 : base(fixture, output, false, true, true, LoggingFramework.NLog)
             {